[search-tips] project arch edits
