+++ js/search-tips/.gitignore
@@ -1,0 +1,3 @@
+/node_modules/
+/.idea/
+/package-lock.json
