+++ js/search-tips/index.html
@@ -6,6 +6,6 @@
   </head>
   <body>
       <div id="search-page"></div>
-      <script src="build/bundle.js"></script>
+      <script src="bundle.js"></script>
   </body>
 </html>
