--- js/search-tips/reducers/index.js
@@ -1,9 +1,0 @@
-import { combineReducers } from 'redux';
-
-import search from './search-reducer';
-
-const reducer = combineReducers({
-  search
-});
-
-export default reducer;
