--- js/search-tips/reducers/search-reducer.js
@@ -1,11 +1,0 @@
-const search = (state = {}, action) => {
-  switch (action.type) {
-    case 'search':
-      return {};
-    default:
-      return state;
-  }
-
-};
-
-export default search;
