+++ js/search-tips/src/components/search-form/index.js
@@ -1,11 +1,13 @@
 import React from 'react';
-import SearchTips from '../components/search-tips';
+import SearchTips from '../search-tips';
+
+import style from './style.css';
 
 const SearchForm =
   () =>
     (
       <section>
-        <h1>Поисковые подсказки</h1>
+        <h1 className={style.header}>Поисковые подсказки</h1>
         <input className="search-input" type="search"/>
         <SearchTips />
       </section>
