+++ js/search-tips/src/components/search-form/style.css
@@ -1,0 +1,3 @@
+:local(.header) {
+    color: blue;
+}
