+++ js/search-tips/src/components/search-tips/index.js
