+++ js/search-tips/src/containers/search/index.js
@@ -1,8 +1,6 @@
 import { connect } from 'react-redux';
 
-import * as actions from '../actions/search-actions';
-
-import SearchForm from '../components/search-form';
+import SearchForm from '../../components/search-form';
 
 const mapStateToProps = state => ({
