+++ js/search-tips/src/data/words/index.js
