+++ js/search-tips/src/index.js
@@ -2,11 +2,9 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 
-import store from './store/store';
+import store from './store';
 
-import SearchContainer from './containers/search-container';
-
-import './styles/index.css'
+import SearchContainer from './containers/search';
 
 render(
   <Provider store={store}>
