+++ js/search-tips/src/store/search/reducer/index.js
@@ -1,0 +1,11 @@
+const search = (state = {}, action) => {
+    switch (action.type) {
+        case 'search':
+            return {};
+        default:
+            return state;
+    }
+
+};
+
+export default search;
